requirements.txt
paho-mqtt
click
construct
cryptography
zeroconf
attrs
pytz
appdirs
tqdm
netifaces
android_backup
importlib_metadata
croniter
defusedxml
sphinx
sphinx_click
sphinxcontrib-apidoc
sphinx_rtd_theme